+++ src/backend/utils/adt/geo_ops.c
@@ -8,7 +8,7 @@
  *
  *
  * IDENTIFICATION
- *	  $Header: /cvsroot/pgsql/src/backend/utils/adt/geo_ops.c,v 1.51 2000/06/13 07:35:07 tgl Exp $
+ *	  $Header: /cvsroot/pgsql/src/backend/utils/adt/geo_ops.c,v 1.52 2000/06/14 18:17:42 petere Exp $
  *
  *-------------------------------------------------------------------------
  */
@@ -111,7 +111,7 @@ single_decode(char *str, float8 *x, char **s)
 	if (!PointerIsValid(str))
 		return FALSE;
 
-	while (isspace(*str))
+	while (isspace((int) *str))
 		str++;
 	*x = strtod(str, &cp);
 #ifdef GEODEBUG
@@ -119,7 +119,7 @@ single_decode(char *str, float8 *x, char **s)
 #endif
 	if (cp <= str)
 		return FALSE;
-	while (isspace(*cp))
+	while (isspace((int) *cp))
 		cp++;
 
 	if (s != NULL)
@@ -144,33 +144,33 @@ pair_decode(char *str, float8 *x, float8 *y, char **s)
 	if (!PointerIsValid(str))
 		return FALSE;
 
-	while (isspace(*str))
+	while (isspace((int) *str))
 		str++;
 	if ((has_delim = (*str == LDELIM)))
 		str++;
 
-	while (isspace(*str))
+	while (isspace((int) *str))
 		str++;
 	*x = strtod(str, &cp);
 	if (cp <= str)
 		return FALSE;
-	while (isspace(*cp))
+	while (isspace((int) *cp))
 		cp++;
 	if (*cp++ != DELIM)
 		return FALSE;
-	while (isspace(*cp))
+	while (isspace((int) *cp))
 		cp++;
 	*y = strtod(cp, &str);
 	if (str <= cp)
 		return FALSE;
-	while (isspace(*str))
+	while (isspace((int) *str))
 		str++;
 	if (has_delim)
 	{
 		if (*str != RDELIM)
 			return FALSE;
 		str++;
-		while (isspace(*str))
+		while (isspace((int) *str))
 			str++;
 	}
 	if (s != NULL)
@@ -195,7 +195,7 @@ path_decode(int opentype, int npts, char *str, int *isopen, char **ss, Point *p)
 	int			i;
 
 	s = str;
-	while (isspace(*s))
+	while (isspace((int) *s))
 		s++;
 	if ((*isopen = (*s == LDELIM_EP)))
 	{
@@ -204,14 +204,14 @@ path_decode(int opentype, int npts, char *str, int *isopen, char **ss, Point *p)
 			return FALSE;
 		depth++;
 		s++;
-		while (isspace(*s))
+		while (isspace((int) *s))
 			s++;
 
 	}
 	else if (*s == LDELIM)
 	{
 		cp = (s + 1);
-		while (isspace(*cp))
+		while (isspace((int) *cp))
 			cp++;
 		if (*cp == LDELIM)
 		{
@@ -247,7 +247,7 @@ path_decode(int opentype, int npts, char *str, int *isopen, char **ss, Point *p)
 		{
 			depth--;
 			s++;
-			while (isspace(*s))
+			while (isspace((int) *s))
 				s++;
 		}
 		else
@@ -1157,7 +1157,7 @@ path_in(char *str)
 		elog(ERROR, "Bad path external representation '%s'", str);
 
 	s = str;
-	while (isspace(*s))
+	while (isspace((int) *s))
 		s++;
 
 	/* skip single leading paren */
@@ -3845,13 +3845,13 @@ circle_in(char *str)
 	circle = palloc(sizeof(CIRCLE));
 
 	s = str;
-	while (isspace(*s))
+	while (isspace((int) *s))
 		s++;
 	if ((*s == LDELIM_C) || (*s == LDELIM))
 	{
 		depth++;
 		cp = (s + 1);
-		while (isspace(*cp))
+		while (isspace((int) *cp))
 			cp++;
 		if (*cp == LDELIM)
 			s = cp;
@@ -3862,7 +3862,7 @@ circle_in(char *str)
 
 	if (*s == DELIM)
 		s++;
-	while (isspace(*s))
+	while (isspace((int) *s))
 		s++;
 
 	if ((!single_decode(s, &circle->radius, &s)) || (circle->radius < 0))
@@ -3875,7 +3875,7 @@ circle_in(char *str)
 		{
 			depth--;
 			s++;
-			while (isspace(*s))
+			while (isspace((int) *s))
 				s++;
 		}
 		else
